django/my-site/blogapp/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-ckeditor==6.2.0
django-js-asset==2.0.0
mysqlclient==2.1.0
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2021.5
